HACT: remove update_one functionality
