+++ src/etools/applications/partners/tests/test_models.py
@@ -1,4 +1,3 @@
-
 import copy
 import datetime
 from unittest import skip
@@ -11,8 +10,6 @@ from django.utils import six, timezone
 from freezegun import freeze_time
 from mock import Mock, patch
 
-from etools.applications.audit.models import Engagement
-from etools.applications.audit.tests.factories import AuditFactory, SpecialAuditFactory, SpotCheckFactory
 from etools.applications.EquiTrack.tests.cases import BaseTenantTestCase
 from etools.applications.funds.tests.factories import DonorFactory, FundsReservationHeaderFactory, GrantFactory
 from etools.applications.locations.tests.factories import LocationFactory
@@ -21,18 +18,13 @@ from etools.applications.partners.tests.factories import (AgreementAmendmentFact
                                                           AssessmentFactory, FileTypeFactory,
                                                           InterventionAmendmentFactory,
                                                           InterventionAttachmentFactory, InterventionBudgetFactory,
-                                                          InterventionFactory, InterventionPlannedVisitsFactory,
-                                                          InterventionReportingPeriodFactory,
+                                                          InterventionFactory, InterventionReportingPeriodFactory,
                                                           InterventionResultLinkFactory,
                                                           InterventionSectorLocationLinkFactory, PartnerFactory,
                                                           PartnerStaffFactory, PlannedEngagementFactory,
                                                           WorkspaceFileTypeFactory,)
 from etools.applications.reports.tests.factories import (AppliedIndicatorFactory, CountryProgrammeFactory,
                                                          LowerResultFactory, ResultFactory, SectorFactory,)
-from etools.applications.t2f.models import Travel, TravelType
-from etools.applications.t2f.tests.factories import TravelActivityFactory, TravelFactory
-from etools.applications.tpm.models import TPMVisit
-from etools.applications.tpm.tests.factories import TPMActivityFactory, TPMVisitFactory
 from etools.applications.users.tests.factories import UserFactory
 
 
@@ -363,225 +355,6 @@ class TestPartnerOrganizationModel(BaseTenantTestCase):
         self.partner_organization.reported_cy = 510000.00
         self.partner_organization.rating = models.PartnerOrganization.RATING_LOW
         self.assert_min_requirements(2, 1)
-
-    def test_planned_visits_gov(self):
-        self.partner_organization.partner_type = models.PartnerType.GOVERNMENT
-        self.partner_organization.save()
-        intervention = InterventionFactory(
-            agreement=self.pca_signed1,
-            status=models.Intervention.ACTIVE
-        )
-        year = datetime.date.today().year
-        InterventionPlannedVisitsFactory(
-            intervention=intervention,
-            year=year,
-            programmatic_q1=3
-        )
-        InterventionPlannedVisitsFactory(
-            intervention=intervention,
-            year=year - 1,
-            programmatic_q3=2
-        )
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['planned']['total'], 0)
-
-    def test_planned_visits_non_gov(self):
-        self.partner_organization.partner_type = models.PartnerType.UN_AGENCY
-        self.partner_organization.save()
-        intervention = InterventionFactory(
-            agreement=self.pca_signed1,
-            status=models.Intervention.ACTIVE
-        )
-        year = datetime.date.today().year
-        InterventionPlannedVisitsFactory(
-            intervention=intervention,
-            year=year,
-            programmatic_q1=3,
-            programmatic_q4=4,
-        )
-        InterventionPlannedVisitsFactory(
-            intervention=intervention,
-            year=year - 1,
-            programmatic_q2=2
-        )
-        models.PartnerOrganization.planned_visits(
-            self.partner_organization
-        )
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['planned']['total'], 7)
-
-    def test_planned_visits_non_gov_no_pv_intervention(self):
-        self.partner_organization.partner_type = models.PartnerType.UN_AGENCY
-        self.partner_organization.save()
-        intervention1 = InterventionFactory(
-            agreement=self.pca_signed1,
-            status=models.Intervention.ACTIVE
-        )
-        intervention2 = InterventionFactory(
-            agreement=self.pca_signed1,
-            status=models.Intervention.ACTIVE
-        )
-        year = datetime.date.today().year
-        InterventionPlannedVisitsFactory(
-            intervention=intervention1,
-            year=year,
-            programmatic_q1=1,
-            programmatic_q3=3,
-        )
-        InterventionPlannedVisitsFactory(
-            intervention=intervention2,
-            year=year - 1,
-            programmatic_q4=2
-        )
-        models.PartnerOrganization.planned_visits(
-            self.partner_organization
-        )
-        self.assertEqual(
-            self.partner_organization.hact_values['programmatic_visits']['planned']['total'],
-            4
-        )
-
-    def test_programmatic_visits_update_one(self):
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['total'], 0)
-        models.PartnerOrganization.programmatic_visits(
-            self.partner_organization,
-            event_date=datetime.datetime(2013, 5, 26),
-            update_one=True
-        )
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['total'], 1)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q1'], 0)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q2'], 1)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q3'], 0)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q4'], 0)
-
-    def test_programmatic_visits_update_travel_activity(self):
-        tz = timezone.get_default_timezone()
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['total'], 0)
-        traveller = UserFactory()
-        travel = TravelFactory(
-            traveler=traveller,
-            status=Travel.COMPLETED,
-            end_date=datetime.datetime(datetime.datetime.today().year, 9, 1, tzinfo=tz)
-        )
-        TravelActivityFactory(
-            travels=[travel],
-            primary_traveler=traveller,
-            travel_type=TravelType.PROGRAMME_MONITORING,
-            partner=self.partner_organization,
-        )
-        models.PartnerOrganization.programmatic_visits(self.partner_organization)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['total'], 1)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q1'], 0)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q2'], 0)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q3'], 1)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q4'], 0)
-
-    def test_programmatic_visits_update_tpm_visit(self):
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['total'], 0)
-        visit = TPMVisitFactory(
-            status=TPMVisit.UNICEF_APPROVED,
-            date_of_unicef_approved=datetime.datetime(datetime.datetime.today().year, 5, 1)
-        )
-        visit2 = TPMVisitFactory(
-            status=TPMVisit.UNICEF_APPROVED,
-            date_of_unicef_approved=datetime.datetime(datetime.datetime.today().year, 5, 20)
-        )
-        TPMActivityFactory(
-            tpm_visit=visit,
-            partner=self.partner_organization,
-        )
-        TPMActivityFactory(
-            tpm_visit=visit,
-            partner=self.partner_organization,
-        )
-        TPMActivityFactory(
-            tpm_visit=visit2,
-            partner=self.partner_organization,
-        )
-
-        models.PartnerOrganization.programmatic_visits(self.partner_organization)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['total'], 1)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q1'], 0)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q2'], 1)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q3'], 0)
-        self.assertEqual(self.partner_organization.hact_values['programmatic_visits']['completed']['q4'], 0)
-
-    @freeze_time("2013-12-26")
-    def test_spot_checks_update_one(self):
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['total'], 0)
-        models.PartnerOrganization.spot_checks(
-            self.partner_organization,
-            update_one=True,
-        )
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['total'], 1)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q1'], 0)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q2'], 0)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q3'], 0)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q4'], 1)
-
-    @freeze_time("2013-12-26")
-    def test_spot_checks_update_one_with_date(self):
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['total'], 0)
-        models.PartnerOrganization.spot_checks(
-            self.partner_organization,
-            update_one=True,
-            event_date=datetime.datetime(2013, 5, 12)
-        )
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['total'], 1)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q1'], 0)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q2'], 1)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q3'], 0)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q4'], 0)
-
-    def test_spot_checks_update_travel_activity(self):
-        tz = timezone.get_default_timezone()
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['total'], 0)
-        traveller = UserFactory()
-        travel = TravelFactory(
-            traveler=traveller,
-            status=Travel.COMPLETED,
-            completed_at=datetime.datetime(datetime.datetime.today().year, 9, 1, tzinfo=tz)
-        )
-        TravelActivityFactory(
-            travels=[travel],
-            primary_traveler=traveller,
-            travel_type=TravelType.SPOT_CHECK,
-            partner=self.partner_organization,
-        )
-
-        SpotCheckFactory(
-            partner=self.partner_organization,
-            status=Engagement.FINAL,
-            date_of_draft_report_to_unicef=datetime.datetime(datetime.datetime.today().year, 4, 1)
-        )
-        models.PartnerOrganization.spot_checks(self.partner_organization)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['total'], 2)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q1'], 0)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q2'], 1)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q3'], 1)
-        self.assertEqual(self.partner_organization.hact_values['spot_checks']['completed']['q4'], 0)
-
-    @freeze_time("2013-12-26")
-    def test_audits_completed_update_one(self):
-        self.assertEqual(self.partner_organization.hact_values['audits']['completed'], 0)
-        models.PartnerOrganization.audits_completed(
-            self.partner_organization,
-            update_one=True,
-        )
-        self.assertEqual(self.partner_organization.hact_values['audits']['completed'], 1)
-
-    def test_audits_completed_update_travel_activity(self):
-        self.assertEqual(self.partner_organization.hact_values['audits']['completed'], 0)
-        AuditFactory(
-            partner=self.partner_organization,
-            status=Engagement.FINAL,
-            date_of_draft_report_to_unicef=datetime.datetime(datetime.datetime.today().year, 4, 1)
-        )
-        SpecialAuditFactory(
-            partner=self.partner_organization,
-            status=Engagement.FINAL,
-            date_of_draft_report_to_unicef=datetime.datetime(datetime.datetime.today().year, 8, 1)
-        )
-        models.PartnerOrganization.audits_completed(self.partner_organization)
-        self.assertEqual(self.partner_organization.hact_values['audits']['completed'], 2)
 
     def test_partner_organization_get_admin_url(self):
         "Test that get_admin_url produces the URL we expect."
