+++ src/etools/applications/partners/views/interventions_v2.py
@@ -27,16 +27,11 @@ from etools.applications.partners.models import (Intervention, InterventionAmend
                                                  InterventionPlannedVisits, InterventionReportingPeriod,
                                                  InterventionResultLink, InterventionSectorLocationLink,)
 from etools.applications.partners.permissions import PartnershipManagerPermission, PartnershipManagerRepPermission
-from etools.applications.partners.serializers.exports.interventions import (InterventionAmendmentExportFlatSerializer,
-                                                                            InterventionAmendmentExportSerializer,
-                                                                            InterventionExportFlatSerializer,
-                                                                            InterventionExportSerializer,
-                                                                            InterventionIndicatorExportFlatSerializer,
-                                                                            InterventionIndicatorExportSerializer,
-                                                                            InterventionResultExportFlatSerializer,
-                                                                            InterventionResultExportSerializer,
-                                                                            InterventionSectorLocationLinkExportFlatSerializer,
-                                                                            InterventionSectorLocationLinkExportSerializer,)
+from etools.applications.partners.serializers.exports.interventions import (
+    InterventionAmendmentExportFlatSerializer, InterventionAmendmentExportSerializer, InterventionExportFlatSerializer,
+    InterventionExportSerializer, InterventionIndicatorExportFlatSerializer, InterventionIndicatorExportSerializer,
+    InterventionResultExportFlatSerializer, InterventionResultExportSerializer,
+    InterventionSectorLocationLinkExportFlatSerializer, InterventionSectorLocationLinkExportSerializer,)
 from etools.applications.partners.serializers.interventions_v2 import (InterventionAmendmentCUSerializer,
                                                                        InterventionAttachmentSerializer,
                                                                        InterventionBudgetCUSerializer,
