+++ src/etools/applications/vision/tests/test_tasks.py
@@ -1,5 +1,4 @@
 # Python imports
-
 import datetime
 
 from django.test import SimpleTestCase
